Created a 'run' method. This makes the code more readable.
Also added output_template and track_output keywords. If the 
environmental parameter VASP_COMMAND is used to execute VASP,
the stdout from VASP is now written to a file (default vasp.out) 
instead of to the stdout of ASE. If VASP_SCRIPT is used the VASP stdout, 
is still printed as stdout for the ASE run (can this be done also for 
VASP_SCRIPT?).



git-svn-id: https://svn.fysik.dtu.dk/projects/ase/trunk@1258 5af997e4-4c3d-0410-961e-ac0512e437d9